requirements.txt
streamlit
pandas
openpyxl
xlsxwriter
azure-ai-documentintelligence